requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.43.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
boto3==1.35.99
botocore==1.35.99
cachetools==5.5.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
Django==5.1.4
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
dogesec_commons @ https://github.com/muchdogesec/dogesec_commons/releases/download/main-2024-12-23-10-31-09/dogesec_commons-0.0.3-py3-none-any.whl#sha256=33d7832a6b0a51c0039df1f0b9fd975900de2c61b7f3c2a3d031a1952e7da352
drf-spectacular==0.28.0
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.12.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
llama-index-core==0.12.10.post1
llama-index-llms-anthropic==0.6.3
llama-index-llms-gemini==0.4.3
llama-index-llms-openai==0.3.13
marshmallow==3.25.1
mistune==3.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.1
openai==1.59.7
packaging==24.2
pillow==10.4.0
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
python-arango==8.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.22.3
rsa==4.9
s3transfer==0.10.4
setuptools==75.8.0
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlparse==0.5.3
stix2==3.0.1
stix2-patterns==2.0.0
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
https://github.com/muchdogesec/txt2detection/releases/download/v0.0.1/txt2detection-0.0.1-py3-none-any.whl
https://github.com/muchdogesec/stix2arango/releases/download/main-2025-01-14-11-49-13/stix2arango-0.0.3-py3-none-any.whl
https://github.com/muchdogesec/file2txt/releases/download/main-2024-11-29-15-39-43/file2txt-0.0.1b2-py3-none-any.whl
